fix: fall back to looser parsing if necessary (#68)
This is a simple fix: if parsing a JS file fails, fall back to trying to parse as a script explicitly and allowing returns outside functions.

The came up from trying to use the official mocha package in our test suite. The build of the app fails because `mkdirp` has a `bin\cmd.js` file using a return "globally" which is allowed under Node.

I think there's a larger discussion to be had as to whether we should be parsing/transpiling/minifying 3rd party node_modules, but for now I think having a fall back like this makes some sense.