Remove async transform explicit plugin listing (#77)
This is causing master builds to fail suddenly on the SDK. It appears to be from a. fluke in how the node_modules folders are getting constructed that this particular plugin lives underneath preset-env and resolution from our jsanalyze.js file fails. In a local test, it appears we don't need to manually add this plugin to the list and it just gets run when the targeted platform needs it (via preset-env).