fix(jsanalyze): handling input source maps, source map handling on iOS
- add sourceMappingURL data: comment ourselves, appending extra newlien to work around iOS require issue
- support consuming input inline data: source maps and external file: source maps
- set sourceRoot in final sourcemap, delete sourcesContent on Android
- On iOS, if source basename matches generated file basename, just report basename as source, not a path
or else Safari/Web Inspector will not show the file. This can lead to original source files
not being shown at all if their basename matches that of another source.
It also means the source entries in iOS source maps may be "wrong" in that they strip the full path.
Consumers trying to resolve back to original filepaths will need to have some knowledge of how we copy files.
The only way to get around this is to try and avoid matching basenames between multiple sources and between source
and generated filename.

Fixes TIMOB-27098